+++ src/trap-caching.ts
@@ -1,14 +1,16 @@
 import * as fs from "fs";
 import * as path from "path";
 
-import * as cache from "@actions/cache";
+import * as actionsCache from "@actions/cache";
 
 import * as actionsUtil from "./actions-util";
+import * as apiClient from "./api-client";
 import { CodeQL } from "./codeql";
 import type { Config } from "./config-utils";
+import { Feature, FeatureEnablement } from "./feature-flags";
 import { Language } from "./languages";
 import { Logger } from "./logging";
-import { tryGetFolderBytes, withTimeout } from "./util";
+import { isHTTPError, tryGetFolderBytes, withTimeout, wrapError } from "./util";
 
 // This constant should be bumped if we make a breaking change
 // to how the CodeQL Action stores or retrieves the TRAP cache,
@@ -16,6 +18,8 @@ import { tryGetFolderBytes, withTimeout } from "./util";
 // this for CLI/extractor changes, since the CLI version also
 // goes into the cache key.
 const CACHE_VERSION = 1;
+
+const CODEQL_TRAP_CACHE_PREFIX = "codeql-trap";
 
 // This constant sets the minimum size in megabytes of a TRAP
 // cache for us to consider it worth uploading.
@@ -87,7 +91,7 @@ export async function downloadTrapCaches(
     );
     const found = await withTimeout(
       MAX_CACHE_OPERATION_MS,
-      cache.restoreCache([cacheDir], preferredKey, [
+      actionsCache.restoreCache([cacheDir], preferredKey, [
         // Fall back to any cache with the right key prefix
         await cachePrefix(codeql, language),
       ]),
@@ -147,7 +151,7 @@ export async function uploadTrapCaches(
     logger.info(`Uploading TRAP cache to Actions cache with key ${key}`);
     await withTimeout(
       MAX_CACHE_OPERATION_MS,
-      cache.saveCache([cacheDir], key),
+      actionsCache.saveCache([cacheDir], key),
       () => {
         logger.info(
           `Timed out waiting for TRAP cache for ${language} to upload, will continue without uploading`,
@@ -156,6 +160,107 @@ export async function uploadTrapCaches(
     );
   }
   return true;
+}
+
+export interface TrapCacheCleanupStatusReport {
+  trap_cache_cleanup_error?: string;
+  trap_cache_cleanup_size_bytes?: number;
+  trap_cache_cleanup_skipped_because?: string;
+}
+
+export async function cleanupTrapCaches(
+  config: Config,
+  features: FeatureEnablement,
+  logger: Logger,
+): Promise<TrapCacheCleanupStatusReport> {
+  if (!(await features.getValue(Feature.CleanupTrapCaches))) {
+    return {
+      trap_cache_cleanup_skipped_because: "feature disabled",
+    };
+  }
+  if (!(await actionsUtil.isAnalyzingDefaultBranch())) {
+    return {
+      trap_cache_cleanup_skipped_because: "not analyzing default branch",
+    };
+  }
+
+  try {
+    let totalBytesCleanedUp = 0;
+
+    const allCaches = await apiClient.listActionsCaches(
+      CODEQL_TRAP_CACHE_PREFIX,
+      await actionsUtil.getRef(),
+    );
+
+    for (const language of config.languages) {
+      if (config.trapCaches[language]) {
+        const cachesToRemove = await getTrapCachesForLanguage(
+          allCaches,
+          language,
+          logger,
+        );
+        // Dates returned by the API are in ISO 8601 format, so we can sort them lexicographically
+        cachesToRemove.sort((a, b) => a.created_at.localeCompare(b.created_at));
+        // Keep the most recent cache
+        const mostRecentCache = cachesToRemove.pop();
+        logger.debug(
+          `Keeping most recent TRAP cache (${JSON.stringify(mostRecentCache)})`,
+        );
+
+        if (cachesToRemove.length === 0) {
+          logger.info(`No TRAP caches to clean up for ${language}.`);
+          continue;
+        }
+
+        for (const cache of cachesToRemove) {
+          logger.debug(`Cleaning up TRAP cache (${JSON.stringify(cache)})`);
+          await apiClient.deleteActionsCache(cache.id);
+        }
+        const bytesCleanedUp = cachesToRemove.reduce(
+          (acc, item) => acc + item.size_in_bytes,
+          0,
+        );
+        totalBytesCleanedUp += bytesCleanedUp;
+        const megabytesCleanedUp = (bytesCleanedUp / (1024 * 1024)).toFixed(2);
+        logger.info(
+          `Cleaned up ${megabytesCleanedUp} MiB of old TRAP caches for ${language}.`,
+        );
+      }
+    }
+    return { trap_cache_cleanup_size_bytes: totalBytesCleanedUp };
+  } catch (e) {
+    if (isHTTPError(e) && e.status === 403) {
+      logger.warning(
+        "Could not cleanup TRAP caches as the token did not have the required permissions. " +
+          'To clean up TRAP caches, ensure the token has the "actions:write" permission. ' +
+          "For more information, see https://docs.github.com/en/actions/using-jobs/assigning-permissions-to-jobs",
+      );
+    } else {
+      logger.info(`Failed to cleanup TRAP caches, continuing. Details: ${e}`);
+    }
+    return { trap_cache_cleanup_error: wrapError(e).message };
+  }
+}
+
+async function getTrapCachesForLanguage(
+  allCaches: apiClient.ActionsCacheItem[],
+  language: Language,
+  logger: Logger,
+): Promise<Array<Required<apiClient.ActionsCacheItem>>> {
+  logger.debug(`Listing TRAP caches for ${language}`);
+
+  for (const cache of allCaches) {
+    if (!cache.created_at || !cache.id || !cache.key || !cache.size_in_bytes) {
+      throw new Error(
+        "An unexpected cache item was returned from the API that was missing one or " +
+          `more required fields: ${JSON.stringify(cache)}`,
+      );
+    }
+  }
+
+  return allCaches.filter((cache) => {
+    return cache.key?.includes(`-${language}-`);
+  }) as Array<Required<apiClient.ActionsCacheItem>>;
 }
 
 export async function getLanguagesSupportingCaching(
@@ -225,7 +330,7 @@ async function cachePrefix(
   codeql: CodeQL,
   language: Language,
 ): Promise<string> {
-  return `codeql-trap-${CACHE_VERSION}-${
+  return `${CODEQL_TRAP_CACHE_PREFIX}-${CACHE_VERSION}-${
     (await codeql.getVersion()).version
   }-${language}-`;
 }